requirements.txt
# Core dependencies
networkx>=3.0
pandas>=2.0
scipy>=1.11.0

# AI & Semantic Analysis
google-genai>=0.1.0
python-dotenv>=1.0.0

# Visualization
plotly>=5.0
matplotlib>=3.8.0
pyvis>=0.3.0

# Testing
pytest>=7.0
pytest-cov>=4.0
hypothesis>=6.0

# Type checking
mypy>=1.0
types-requests
